Reformat workspace

+++ include/core/math/shape/Simplex.h
@@ -3,13 +3,13 @@
 #include <algorithm>
 #include <array>
 #include <initializer_list>	
-#include "math/vector.h"
+#include "core/math/vector.h"
 
 
 struct Simplex {
 private:
 	std::array<vec3f, 4> pts;
-	unsigned int _size;
+	size_t _size;
 public:
 	Simplex() : pts({ vec3f(0,0,0), vec3f(0,0,0), vec3f(0,0,0), vec3f(0,0,0) }), _size(0) { }
 	Simplex& operator=(std::initializer_list<vec3f> list) {
@@ -23,7 +23,7 @@ public:
 
 	void push(vec3f pt) {
 		pts = { pt, pts[0], pts[1], pts[2] };
-		_size = std::min(_size + 1, 4u);
+		_size = std::min(_size + 1, 4ull);
 	}
 
 	vec3f& operator[](unsigned int i) { return this->pts[i]; }
